Improvement of cell editable
To avoid in case that the user click cell A once then go to click other cell. It will cause cell A to be set to contentEditable="true" after first click. If the user come back to edit cell A, he will find that it is editable after the NEW first click.